Efficient Leaf Struct
Turned the leaf struct into an array which improves efficiency, readability, and simplicity, etc.
writing more to notepad and getting weird error in the Selection enumerator (that controls the user custom bearing) which returns a default (which shouldnt happen) .
Note: Can pass class pointers to function call which is useful, see GetDoubleLapLoad() on MasonryChecker.cpp Ln 323 & 445
Weird error on line 200 after converting Load array to a matrix. TODO -> FIX
Started to look into getting the number of leaves so that in future a single leaf analysis can be carried out, i.e. add all load to 1 leaf not 2 and calculate eccentric loads from both sides of the wall.